extract Activity functionality to a hook; todo: on prop drilling indentation
